+++ c-api/src/lib.rs
@@ -6,11 +6,6 @@ use std::{ptr, slice, str};
 use thiserror::Error;
 
 #[inline]
-fn to_ptr<T>(val: T) -> *const T {
-    Box::into_raw(Box::new(val))
-}
-
-#[inline]
 fn to_ptr_mut<T>(val: T) -> *mut T {
     Box::into_raw(Box::new(val))
 }